% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom.R
\name{multi_batch}
\alias{multi_batch}
\title{Multi batch}
\usage{
multi_batch(batch_list, batch_func, source_desc)
}
\arguments{
\item{batch_list}{}

\item{batch_func}{}

\item{source_desc}{}
}
\description{
Applies a batch method over a list of multiple batches
}
