% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom.R
\name{multi_contacts}
\alias{multi_contacts}
\title{Multi contacts}
\usage{
multi_contacts(lid, n = v6_countContacts(lid), size = 1000, returned_fields,
  ...)
}
\arguments{
\item{lid}{}

\item{n}{}

\item{size}{}

\item{returned_fields}{}

\item{...}{}
}
\description{
A wrapper for v6_searchContacts where the numer of contacts is large.
}
