% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom.R
\name{multi_edit}
\alias{multi_edit}
\title{Multi edit}
\usage{
multi_edit(list_id, contacts, chunk_size = 5000, ...)
}
\arguments{
\item{list_id}{}

\item{contacts}{}

\item{chunk_size}{}

\item{...}{}
}
\description{
A wrapper for v6_editContacts for large number contacts to edit
}
