% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/autoresponder.R
\name{v6_addAutoresponder}
\alias{v6_addAutoresponder}
\title{v6_addAutoresponder}
\usage{
v6_addAutoresponder()
}
\description{
v6_addAutoresponder
}
\seealso{
\url{http://developers.vision6.com.au/3.0/method/addAutoresponder}
}

