% Generated by roxygen2 (4.0.2): do not edit by hand
\name{v6_addBatchByContactSearch}
\alias{v6_addBatchByContactSearch}
\title{v6_addBatchByContactSearch}
\usage{
v6_addBatchByContactSearch()
}
\description{
v6_addBatchByContactSearch
}
\seealso{
\url{http://developers.vision6.com.au/3.0/method/addBatchByContactSearch}
}

