% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/contact.R
\name{v6_addContacts}
\alias{v6_addContacts}
\title{v6_addContacts}
\usage{
v6_addContacts(list_id, contacts, overwrite = FALSE,
  remove_unsubscribers = 0)
}
\description{
v6_addContacts
}
\seealso{
\url{http://developers.vision6.com.au/3.0/method/addContacts}
}

