% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/field.R
\name{v6_addField}
\alias{v6_addField}
\title{Adds a new Field to a List.}
\usage{
v6_addField(list_id, field_details)
}
\arguments{
\item{list_id}{(int) The ID of the List to add the Field to.}

\item{field_details}{(array) A list or data frame of Field details.
See \\code{\link{v6_fieldDetails}} for a description of valid values.}
}
\value{
the ID of the newly added Field on success.
}
\description{
Adds a new Field to a List.
}
\seealso{
\url{http://developers.vision6.com.au/3.0/method/addField}
}
