% Generated by roxygen2 (4.0.2): do not edit by hand
\name{v6_addFolder}
\alias{v6_addFolder}
\title{Creates a new Folder.}
\usage{
v6_addFolder(folder_details)
}
\arguments{
\item{folder_details}{(array) A list or data frame of Field details.
See \\code{\link{v6_folderDetails}} for a description of valid values.}
}
\value{
the ID of the newly added Folder on success.
}
\description{
Creates a new Folder.
}
\seealso{
\url{http://developers.vision6.com.au/3.0/method/addFolder}
}

