% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list.R
\name{v6_addList}
\alias{v6_addList}
\title{Creates a new Contact List.}
\usage{
v6_addList(list_details)
}
\arguments{
\item{list_details}{(array) A list or data frame of List details.
See \\code{\link{v6_listDetails}} for a description of valid values.}
}
\value{
the ID of the newly created List on success.
}
\description{
Creates a new Contact List.
}
\seealso{
\url{http://developers.vision6.com.au/3.0/method/addList}
}
