% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/contact.R
\name{v6_confirmContact}
\alias{v6_confirmContact}
\title{v6_confirmContact}
\usage{
v6_confirmContact()
}
\description{
v6_confirmContact
}
\seealso{
\url{http://developers.vision6.com.au/3.0/method/confirmContact}
}

