% Generated by roxygen2 (4.0.2): do not edit by hand
\name{v6_countAutoresponders}
\alias{v6_countAutoresponders}
\title{v6_countAutoresponders}
\usage{
v6_countAutoresponders()
}
\description{
v6_countAutoresponders
}
\seealso{
\url{http://developers.vision6.com.au/3.0/method/countAutoresponders}
}

