% Generated by roxygen2 (4.0.2): do not edit by hand
\name{v6_countBatches}
\alias{v6_countBatches}
\title{Counts the number of Batches that match the search criteria.}
\usage{
v6_countBatches(search_criteria = "[]")
}
\arguments{
\item{search_criteria}{(array - optional) An array of search criteria that
selects the Batches to be counted. See \\code{\link{v6_batchDetails}}
for a description of the values that can be searched on.}
}
\value{
The number of Batches that match the search criteria.
}
\description{
Counts the number of Batches that match the search criteria.
}
\seealso{
\url{http://developers.vision6.com.au/3.0/method/countBatches}
}

