% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/contact.R
\name{v6_countContacts}
\alias{v6_countContacts}
\title{v6_countContacts}
\usage{
v6_countContacts()
}
\description{
v6_countContacts
}
\seealso{
\url{http://developers.vision6.com.au/3.0/method/countContacts}
}

