% Generated by roxygen2 (4.0.2): do not edit by hand
\name{v6_countFields}
\alias{v6_countFields}
\title{Counts the number of Fields in a List that match the search criteria.}
\usage{
v6_countFields(list_id, search_criteria = "[]")
}
\arguments{
\item{list_id}{(int) ID of the List containing the Fields.}

\item{search_criteria}{(array - optional) An array of search criteria that
selects the Fields to be counted. See \\code{\link{v6_fieldDetails}} for
a description of the values that can be searched on.}
}
\value{
The number of Fields in the List that match the search criteria.
}
\description{
Counts the number of Fields in a List that match the search criteria.
}
\seealso{
\url{http://developers.vision6.com.au/3.0/method/countFields}
}

