% Generated by roxygen2 (4.0.2): do not edit by hand
\name{v6_countFiles}
\alias{v6_countFiles}
\title{Counts the number of Files that match the search criteria.}
\usage{
v6_countFiles(search_criteria = "[]")
}
\arguments{
\item{search_criteria}{(array - optional) An array of search criteria that
selects the Files to be counted. See \\code{\link{v6_FileDetails}} for a
description of the values that can be searched on.}
}
\value{
The number of Files that match the search criteria.
}
\description{
Counts the number of Files that match the search criteria.
}
\seealso{
\url{http://developers.vision6.com.au/3.0/method/countFiles}
}

