% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list.R
\name{v6_countLists}
\alias{v6_countLists}
\title{Counts the number of Lists that match the search criteria.}
\usage{
v6_countLists(search_criteria = "[]")
}
\arguments{
\item{search_criteria}{(array - optional) An array of search criteria that
selects the Lists to be counted. See \\code{\link{v6_listDetails}} for a
description of the values that can be searched on.}
}
\value{
The number of Lists that match the search criteria.
}
\description{
Counts the number of Lists that match the search criteria.
}
\seealso{
\url{http://developers.vision6.com.au/3.0/method/countLists}
}
