% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/message.R
\name{v6_countMessages}
\alias{v6_countMessages}
\title{v6_countMessages}
\usage{
v6_countMessages()
}
\description{
v6_countMessages
}
\seealso{
\url{http://developers.vision6.com.au/3.0/method/countMessages}
}
