% Generated by roxygen2 (4.0.2): do not edit by hand
\name{v6_deleteAutoresponder}
\alias{v6_deleteAutoresponder}
\title{v6_deleteAutoresponder}
\usage{
v6_deleteAutoresponder()
}
\description{
v6_deleteAutoresponder
}
\seealso{
\url{http://developers.vision6.com.au/3.0/method/deleteAutoresponder}
}

