% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contact.R
\name{v6_deleteContacts}
\alias{v6_deleteContacts}
\title{v6_deleteContacts}
\usage{
v6_deleteContacts(list_id, contact_id)
}
\description{
v6_deleteContacts
}
\seealso{
\url{http://developers.vision6.com.au/3.0/method/deleteContacts}
}
