% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/field.R
\name{v6_deleteField}
\alias{v6_deleteField}
\title{Counts the number of Fields in a List that match the search criteria.}
\usage{
v6_deleteField(list_id, field_id)
}
\arguments{
\item{list_id}{(int) ID of the List containing the Field.}

\item{field_id}{(int) ID of the Field to delete.}
}
\value{
The number of Fields in the List that match the search criteria.
}
\description{
Counts the number of Fields in a List that match the search criteria.
}
\seealso{
\url{http://developers.vision6.com.au/3.0/method/deleteField}
}

