% Generated by roxygen2 (4.0.2): do not edit by hand
\name{v6_deleteFile}
\alias{v6_deleteFile}
\title{Deletes the specified File.}
\usage{
v6_deleteFile(file_id)
}
\arguments{
\item{file_id}{(int) The ID of the File to delete.}
}
\value{
true on success.
}
\description{
Deletes the specified File.
}
\seealso{
\url{http://developers.vision6.com.au/3.0/method/deleteFile}
}

