% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list.R
\name{v6_deleteList}
\alias{v6_deleteList}
\title{Deletes a List and all of its Contacts.}
\usage{
v6_deleteList(list_id)
}
\arguments{
\item{list_id}{(int) The ID of the List to delete.}
}
\value{
true on success.
}
\description{
Deletes a List and all of its Contacts.
}
\seealso{
\url{http://developers.vision6.com.au/3.0/method/deleteList}
}
