% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/message.R
\name{v6_deleteMessage}
\alias{v6_deleteMessage}
\title{v6_deleteMessage}
\usage{
v6_deleteMessage()
}
\description{
v6_deleteMessage
}
\seealso{
\url{http://developers.vision6.com.au/3.0/method/deleteMessage}
}
