% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/contact.R
\name{v6_editContacts}
\alias{v6_editContacts}
\title{v6_editContacts}
\usage{
v6_editContacts(list_id, contacts, trigger_update_profile = FALSE)
}
\description{
v6_editContacts
}
\seealso{
\url{http://developers.vision6.com.au/3.0/method/editContacts}
}

