% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/field.R
\name{v6_editField}
\alias{v6_editField}
\title{Edits a Field in a List.}
\usage{
v6_editField(list_id, field_details)
}
\arguments{
\item{list_id}{(int) The ID of the List containing the Field.}

\item{field_details}{(array) A list or data frame of Field details to be edited.
See \\code{\link{v6_fieldDetails}} for a description of valid values.}
}
\value{
the ID of the edited Field on success.
}
\description{
Edits a Field in a List.
}
\seealso{
\url{http://developers.vision6.com.au/3.0/method/editField}
}

