% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/folder.R
\name{v6_editFolder}
\alias{v6_editFolder}
\title{Renames the specified Folder.}
\usage{
v6_editFolder(folder_details)
}
\arguments{
\item{folder_details}{(array) A list or data frame of Folder details to be edited.
See \\code{\link{v6_folderDetails}} for a description of valid values.}
}
\value{
true on success.
}
\description{
Renames the specified Folder.
}
\seealso{
\url{http://developers.vision6.com.au/3.0/method/editFolder}
}

