% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/list.R
\name{v6_editList}
\alias{v6_editList}
\title{Edits a List.}
\usage{
v6_editList(list_details)
}
\arguments{
\item{list_details}{(array) A list or data frame of List details to be edited.
See \\code{\link{v6_listDetails}} for a description of valid values.}
}
\value{
the List ID on success.
}
\description{
Edits a List.
}
\seealso{
\url{http://developers.vision6.com.au/3.0/method/editList}
}

