% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/file.R
\name{v6_findFileInFolder}
\alias{v6_findFileInFolder}
\title{Returns the ID of a File with the specified name.}
\usage{
v6_findFileInFolder(file_name, folder_id)
}
\arguments{
\item{file_name}{(string) File name to search for.}

\item{folder_id}{(int) ID of Folder to search in.}
}
\value{
ID of the matched file.
}
\description{
Returns the ID of a File with the specified name.
}
\seealso{
\url{http://developers.vision6.com.au/3.0/method/findFileInFolder}
}

