% Generated by roxygen2 (4.0.2): do not edit by hand
\name{v6_getAllCurrentBatches}
\alias{v6_getAllCurrentBatches}
\title{v6_getAllCurrentBatches}
\usage{
v6_getAllCurrentBatches()
}
\description{
v6_getAllCurrentBatches
}
\seealso{
\url{http://developers.vision6.com.au/3.0/method/getAllCurrentBatches}
}

