% Generated by roxygen2 (4.0.2): do not edit by hand
\name{v6_getAutoresponderBatchId}
\alias{v6_getAutoresponderBatchId}
\title{v6_getAutoresponderBatchId}
\usage{
v6_getAutoresponderBatchId()
}
\description{
v6_getAutoresponderBatchId
}
\seealso{
\url{http://developers.vision6.com.au/3.0/method/getAutoresponderBatchId}
}

