% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch.R
\name{v6_getBatchAutoDeactivated}
\alias{v6_getBatchAutoDeactivated}
\title{v6_getBatchAutoDeactivated}
\usage{
v6_getBatchAutoDeactivated(batch_id, search_criteria = "[]", limit = 0,
  offset = 0, sort_by = "", sort_order = "")
}
\description{
v6_getBatchAutoDeactivated
}
\seealso{
\url{http://developers.vision6.com.au/3.0/method/getBatchAutoDeactivated}
}
