% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch.R
\name{v6_getBatchBounced}
\alias{v6_getBatchBounced}
\title{v6_getBatchBounced}
\usage{
v6_getBatchBounced(batch_id, search_criteria = "[]", limit = 0,
  offset = 0, sort_by = "", sort_order = "")
}
\description{
v6_getBatchBounced
}
\seealso{
\url{http://developers.vision6.com.au/3.0/method/getBatchBounced}
}
