% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch.R
\name{v6_getBatchById}
\alias{v6_getBatchById}
\title{v6_getBatchById}
\usage{
v6_getBatchById(batch_id)
}
\description{
v6_getBatchById
}
\seealso{
\url{http://developers.vision6.com.au/3.0/method/getBatchById}
}
