% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch.R
\name{v6_getBatchComplained}
\alias{v6_getBatchComplained}
\title{v6_getBatchComplained}
\usage{
v6_getBatchComplained(batch_id, search_criteria = "[]", limit = 0,
  offset = 0, sort_by = "", sort_order = "")
}
\description{
v6_getBatchComplained
}
\seealso{
\url{http://developers.vision6.com.au/3.0/method/getBatchComplained}
}
