% Generated by roxygen2 (4.0.2): do not edit by hand
\name{v6_getBatchIdByQueueId}
\alias{v6_getBatchIdByQueueId}
\title{v6_getBatchIdByQueueId}
\usage{
v6_getBatchIdByQueueId()
}
\description{
v6_getBatchIdByQueueId
}
\seealso{
\url{http://developers.vision6.com.au/3.0/method/getBatchIdByQueueId}
}

