% Generated by roxygen2 (4.0.2): do not edit by hand
\name{v6_getBatchLinkStatistics}
\alias{v6_getBatchLinkStatistics}
\title{v6_getBatchLinkStatistics}
\usage{
v6_getBatchLinkStatistics(batch_id)
}
\description{
v6_getBatchLinkStatistics
}
\seealso{
\url{http://developers.vision6.com.au/3.0/method/getBatchLinkStatistics}
}

