% Generated by roxygen2 (4.0.2): do not edit by hand
\name{v6_getBatchNotOpened}
\alias{v6_getBatchNotOpened}
\title{v6_getBatchNotOpened}
\usage{
v6_getBatchNotOpened(batch_id, search_criteria = "[]", limit = 0,
  offset = 0, sort_by = "", sort_order = "")
}
\description{
v6_getBatchNotOpened
}
\seealso{
\url{http://developers.vision6.com.au/3.0/method/getBatchNotOpened}
}

