% Generated by roxygen2 (4.0.2): do not edit by hand
\name{v6_getBatchOpened}
\alias{v6_getBatchOpened}
\title{v6_getBatchOpened}
\usage{
v6_getBatchOpened(batch_id, search_criteria = "[]", limit = 0, offset = 0,
  sort_by = "", sort_order = "")
}
\description{
v6_getBatchOpened
}
\seealso{
\url{http://developers.vision6.com.au/3.0/method/getBatchOpened}
}

