% Generated by roxygen2 (4.0.2): do not edit by hand
\name{v6_getBatchResponses}
\alias{v6_getBatchResponses}
\title{v6_getBatchResponses}
\usage{
v6_getBatchResponses(batch_id, search_criteria = "[]", limit = 0,
  offset = 0, sort_by = "", sort_order = "")
}
\description{
v6_getBatchResponses
}
\seealso{
\url{http://developers.vision6.com.au/3.0/method/getBatchResponses}
}

