% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch.R
\name{v6_getBatchStatistics}
\alias{v6_getBatchStatistics}
\title{v6_getBatchStatistics}
\usage{
v6_getBatchStatistics(batch_id)
}
\description{
v6_getBatchStatistics
}
\seealso{
\url{http://developers.vision6.com.au/3.0/method/getBatchStatistics}
}
