% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/batch.R
\name{v6_getBatchStatus}
\alias{v6_getBatchStatus}
\title{v6_getBatchStatus}
\usage{
v6_getBatchStatus(batch_id)
}
\description{
v6_getBatchStatus
}
\seealso{
\url{http://developers.vision6.com.au/3.0/method/getBatchStatus}
}

