% Generated by roxygen2 (4.0.2): do not edit by hand
\name{v6_getBatchUnsubscribed}
\alias{v6_getBatchUnsubscribed}
\title{v6_getBatchUnsubscribed}
\usage{
v6_getBatchUnsubscribed(batch_id, search_criteria = "[]", limit = 0,
  offset = 0, sort_by = "", sort_order = "")
}
\description{
v6_getBatchUnsubscribed
}
\seealso{
\url{http://developers.vision6.com.au/3.0/method/getBatchUnsubscribed}
}

