% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch.R
\name{v6_getBatchUpdated}
\alias{v6_getBatchUpdated}
\title{v6_getBatchUpdated}
\usage{
v6_getBatchUpdated()
}
\description{
v6_getBatchUpdated
}
\seealso{
\url{http://developers.vision6.com.au/3.0/method/getBatchUpdated}
}
