% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/contact.R
\name{v6_getContactById}
\alias{v6_getContactById}
\title{v6_getContactById}
\usage{
v6_getContactById()
}
\description{
v6_getContactById
}
\seealso{
\url{http://developers.vision6.com.au/3.0/method/getContactById}
}

