% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file.R
\name{v6_getFileById}
\alias{v6_getFileById}
\title{Returns the details of a specific File.}
\usage{
v6_getFileById(file_id)
}
\arguments{
\item{file_id}{(int) The File ID}
}
\value{
a data frame containing a single record of File details.
        See \code{\link{v6_fileDetails}} for a description of the values returned.
}
\description{
Returns the details of a specific File.
}
\seealso{
\url{http://developers.vision6.com.au/3.0/method/getFileById}
}
