% Generated by roxygen2 (4.0.2): do not edit by hand
\name{v6_getFolderById}
\alias{v6_getFolderById}
\title{Returns the details of a specific Folder.}
\usage{
v6_getFolderById(folder_id)
}
\arguments{
\item{folder_id}{(int) The ID of the Folder to retrieve.}
}
\value{
a data frame containing a single record of Folder details.
        See \code{\link{v6_folderDetails}} for a description of the values returned.
}
\description{
Returns the details of a specific Folder.
}
\seealso{
\url{http://developers.vision6.com.au/3.0/method/getFolderById}
}

