% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/field.R
\name{v6_getFolderIdForField}
\alias{v6_getFolderIdForField}
\title{Returns the Folder ID for a File type Field in a List.}
\usage{
v6_getFolderIdForField(list_id, field_id)
}
\arguments{
\item{list_id}{(int) The list ID of the Field.}

\item{field_id}{(int) The ID of the Field to get Folder ID for.}
}
\value{
the Folder ID for the Field on success.
}
\description{
Returns the Folder ID for a File type Field in a List.
}
\seealso{
\url{http://developers.vision6.com.au/3.0/method/getFolderIdForField}
}
