% Generated by roxygen2 (4.0.2): do not edit by hand
\name{v6_getListById}
\alias{v6_getListById}
\title{Returns details for a specific List.}
\usage{
v6_getListById(list_id)
}
\arguments{
\item{list_id}{(int) The ID of the List to retrieve.}
}
\value{
a data frame containing a single record of List details.
        See \code{\link{v6_listDetails}} for a description of the values returned.
}
\description{
Returns details for a specific List.
}
\seealso{
\url{http://developers.vision6.com.au/3.0/method/getListById}
}

