% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/message.R
\name{v6_getMessageById}
\alias{v6_getMessageById}
\title{v6_getMessageById}
\usage{
v6_getMessageById(message_id, retrieve_content = TRUE)
}
\description{
v6_getMessageById
}
\seealso{
\url{http://developers.vision6.com.au/3.0/method/getMessageById}
}
