% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/contact.R
\name{v6_reactivateContact}
\alias{v6_reactivateContact}
\title{v6_reactivateContact}
\usage{
v6_reactivateContact(list_id, contact_id)
}
\description{
v6_reactivateContact
}
\seealso{
\url{http://developers.vision6.com.au/3.0/method/reactivateContact}
}

