% Generated by roxygen2 (4.0.2): do not edit by hand
\name{v6_resubscribeContact}
\alias{v6_resubscribeContact}
\title{v6_resubscribeContact}
\usage{
v6_resubscribeContact(list_id, contact_id)
}
\description{
v6_resubscribeContact
}
\seealso{
\url{http://developers.vision6.com.au/3.0/method/resubscribeContact}
}

