% Generated by roxygen2 (4.0.2): do not edit by hand
\name{v6_searchAutoresponders}
\alias{v6_searchAutoresponders}
\title{v6_searchAutoresponders}
\usage{
v6_searchAutoresponders()
}
\description{
v6_searchAutoresponders
}
\seealso{
\url{http://developers.vision6.com.au/3.0/method/searchAutoresponders}
}

