% Generated by roxygen2 (4.0.2): do not edit by hand
\name{v6_searchBatches}
\alias{v6_searchBatches}
\title{v6_searchBatches}
\usage{
v6_searchBatches(search_criteria = "[]", limit = 0, offset = 0,
  sort_by = "", sort_order = "")
}
\description{
v6_searchBatches
}
\seealso{
\url{http://developers.vision6.com.au/3.0/method/searchBatches}
}

