% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contact.R
\name{v6_searchContacts}
\alias{v6_searchContacts}
\title{v6_searchContacts}
\usage{
v6_searchContacts(list_id, search_criteria = "[]", limit = 0, offset = 0,
  sort_by = "", sort_order = "", returned_fields = "all")
}
\description{
v6_searchContacts
}
\seealso{
\url{http://developers.vision6.com.au/3.0/method/searchContacts}
}
