% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file.R
\name{v6_searchFiles}
\alias{v6_searchFiles}
\title{Returns the Files that match the search criteria.}
\usage{
v6_searchFiles(search_criteria = "[]", limit = 0, offset = 0,
  sort_by = "", sort_order = "")
}
\arguments{
\item{search_criteria}{(array - optional)
An array of search criteria. See the remarks section below for the
list of values which are searchable. If you do not specify search
criteria, all Files are returned.}

\item{limit}{(int - optional) Maximum number of items to return.}

\item{offset}{(int - optional) Number of items to skip before beginning to
return results. See Using Search Methods for information about using
the offset parameter with the limit parameter to include pagination
in your results.}

\item{sort_by}{(string - optional) The name of the value to sort the results by.}

\item{sort_order}{(string - optional) The order the results are returned in
(either ascending or descending). The default is descending.
Valid values are:
\itemize{
  \item DESC - Sort from highest to lowest (default).
  \item ASC - Sort from lowest to highest.
}}
}
\value{
a data frame containing zero or more records of File details.
        See \code{\link{v6_fileDetails}} for a description of the values returned.
}
\description{
Returns the Files that match the search criteria.
}
\seealso{
\url{http://developers.vision6.com.au/3.0/method/searchFiles}
}
