% Generated by roxygen2 (4.0.2): do not edit by hand
\name{v6_searchFolders}
\alias{v6_searchFolders}
\title{Returns the Folders that match the search criteria.}
\usage{
v6_searchFolders(folder_type, search_criteria = "[]", limit = 0,
  offset = 0, sort_by = "", sort_order = "")
}
\arguments{
\item{folder_type}{(string) The type of Folders to search for.
Valid values are: list, message or file.}

\item{search_criteria}{(array - optional)
An array of search criteria. See the remarks section below for the
list of values which are searchable. If you do not specify search
criteria, all Folders with the specified folder_type are returned.}

\item{limit}{(int - optional) Maximum number of items to return.}

\item{offset}{(int - optional) Number of items to skip before beginning to
return results. See Using Search Methods for information about using
the offset parameter with the limit parameter to include pagination
in your results.}

\item{sort_by}{(string - optional) The name of the value to sort the results by.}

\item{sort_order}{(string - optional) The order the results are returned in
(either ascending or descending). The default is descending.
Valid values are:
\itemize{
  \item DESC - Sort from highest to lowest (default).
  \item ASC - Sort from lowest to highest.
}}
}
\value{
a data frame containing zero or more records of Field details.
        See \code{\link{v6_fieldDetails}} for a description of the values returned.
}
\description{
Returns the Folders that match the search criteria.
}
\seealso{
\url{http://developers.vision6.com.au/3.0/method/searchFolders}
}

