% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/message.R
\name{v6_searchMessages}
\alias{v6_searchMessages}
\title{v6_searchMessages}
\usage{
v6_searchMessages(search_criteria = "[]", limit = 0, offset = 0,
  sort_by = "", sort_order = "")
}
\description{
v6_searchMessages
}
\seealso{
\url{http://developers.vision6.com.au/3.0/method/searchMessages}
}
