% Generated by roxygen2 (4.0.2): do not edit by hand
\name{v6_searchPreviousUnsubscribers}
\alias{v6_searchPreviousUnsubscribers}
\title{v6_searchPreviousUnsubscribers}
\usage{
v6_searchPreviousUnsubscribers()
}
\description{
v6_searchPreviousUnsubscribers
}
\seealso{
\url{http://developers.vision6.com.au/3.0/method/searchPreviousUnsubscribers}
}

