% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/field.R
\name{v6_setFieldFolderId}
\alias{v6_setFieldFolderId}
\title{Sets the File Folder ID for a File type Field in a List.}
\usage{
v6_setFieldFolderId(list_id, field_id, folder_id)
}
\arguments{
\item{list_id}{(int) The List ID of the Field.}

\item{field_id}{(int) The Field ID.}

\item{folder_id}{(int) The Folder ID for the Field.}
}
\value{
true on success.
}
\description{
Sets the File Folder ID for a File type Field in a List.
}
\seealso{
\url{http://developers.vision6.com.au/3.0/method/setFieldFolderId}
}

