% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/contact.R
\name{v6_subscribeContact}
\alias{v6_subscribeContact}
\title{v6_subscribeContact}
\usage{
v6_subscribeContact()
}
\description{
v6_subscribeContact
}
\seealso{
\url{http://developers.vision6.com.au/3.0/method/subscribeContact}
}

