% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contact.R
\name{v6_unsubscribeContact}
\alias{v6_unsubscribeContact}
\title{v6_unsubscribeContact}
\usage{
v6_unsubscribeContact(list_id, email_address)
}
\description{
v6_unsubscribeContact
}
\seealso{
\url{http://developers.vision6.com.au/3.0/method/unsubscribeContact}
}
