% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contact.R
\name{v6_unsubscribeContactById}
\alias{v6_unsubscribeContactById}
\title{v6_unsubscribeContactById}
\usage{
v6_unsubscribeContactById(list_id, contact_id)
}
\description{
v6_unsubscribeContactById
}
\seealso{
\url{http://developers.vision6.com.au/3.0/method/unsubscribeContactById}
}
