% Generated by roxygen2 (4.0.2): do not edit by hand
\name{v6_uploadBase64FileData}
\alias{v6_uploadBase64FileData}
\title{Uploads a File into the File Manager.}
\usage{
v6_uploadBase64FileData(file_details, folder_id, duplicate_action,
  skip_resize_images = NA, skip_resample_images = NA)
}
\arguments{
\item{file_details}{(array) An associative array with the following values:
\tabular{lll}{
  Key \tab Type \tab Description \cr
  name \tab string \tab The name to use for saving the File once it has been uploaded. \cr
  data \tab string \tab The file data encoded with MIME base64. \cr
}}

\item{folder_id}{(int) The ID of the Folder that will contain the uploaded File.}

\item{duplicate_action}{(string) Action to take if a File with the same name
       as the upload File exists. Valid values are:
\itemize{
  \item increment - create a new File named filename_count.jpg, where
        filename is the name of the original file, and count is an
        incremental index for Files with this filename. The first duplicate
        is named filename_1.jpg.
  \item overwrite - overwrite the existing File with the same name.
}}

\item{skip_resize_images}{(bool - optional) Whether or not to resize small images.
       Valid values are:
\itemize{
  \item true - Do not resize images.
  \item false - Resize images to a maximum value of 800px for width and 800px
        for height. Image ratio is kept during resizing.
}}

\item{skip_resample_images}{(bool - optional) Whether or not to correct for
       lower image quality when resizing. Valid values are:
\itemize{
  \item true - Correct for degraded image quality when resizing.
  \item false - Do not correct for degraded image quality when resizing.
}}
}
\value{
a list on success containing:
\itemize{
  \item the ID of the uploaded File
  \item the URL for the File
}
}
\description{
Uploads a File into the File Manager.
}
\seealso{
\url{http://developers.vision6.com.au/3.0/method/uploadBase64FileData}
}

