% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vision6-package.R
\docType{package}
\name{vision6-package}
\alias{vision6-package}
\alias{vision6}
\title{Vision6 for R}
\description{
Vision6 R
}
\details{
\code{vision6} is an R interface to the Vision6
(\url{http://www.vision6.com.au/}) email system.

Implements all methods included in the Vision6 API
(\url{http://developers.vision6.com.au/}).
}
\author{
Danny Smith \email{danny@gorcha.org}
}
\keyword{package}
