% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/batch.R
\name{v6_addBatchByContactSearch}
\alias{v6_addBatchByContactSearch}
\title{v6_addBatchByContactSearch}
\usage{
v6_addBatchByContactSearch(message_id, lists, search_criteria = "[]",
  send_time = "", is_scheduled = FALSE)
}
\description{
v6_addBatchByContactSearch
}
\seealso{
\url{http://developers.vision6.com.au/3.0/method/addBatchByContactSearch}
}

