% Generated by roxygen2 (4.0.2): do not edit by hand
\name{v6_clearList}
\alias{v6_clearList}
\title{Deletes all Contacts in a List.}
\usage{
v6_clearList(list_id)
}
\arguments{
\item{list_id}{(int) The ID of the List to clear.}
}
\value{
true on success.
}
\description{
Deletes all Contacts in a List.
}
\seealso{
\url{http://developers.vision6.com.au/3.0/method/clearList}
}

