% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contact.R
\name{v6_countContacts}
\alias{v6_countContacts}
\title{v6_countContacts}
\usage{
v6_countContacts(list_id, search_criteria = "[]")
}
\description{
v6_countContacts
}
\seealso{
\url{http://developers.vision6.com.au/3.0/method/countContacts}
}
