% Generated by roxygen2 (4.0.2): do not edit by hand
\name{v6_countPreviousUnsubscribers}
\alias{v6_countPreviousUnsubscribers}
\title{v6_countPreviousUnsubscribers}
\usage{
v6_countPreviousUnsubscribers()
}
\description{
v6_countPreviousUnsubscribers
}
\seealso{
\url{http://developers.vision6.com.au/3.0/method/countPreviousUnsubscribers}
}

