% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contact.R
\name{v6_deactivateContact}
\alias{v6_deactivateContact}
\title{v6_deactivateContact}
\usage{
v6_deactivateContact(list_id, contact_id)
}
\description{
v6_deactivateContact
}
\seealso{
\url{http://developers.vision6.com.au/3.0/method/deactivateContact}
}
