% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/autoresponder.R
\name{v6_deleteAutoresponder}
\alias{v6_deleteAutoresponder}
\title{v6_deleteAutoresponder}
\usage{
v6_deleteAutoresponder()
}
\description{
v6_deleteAutoresponder
}
\seealso{
\url{http://developers.vision6.com.au/3.0/method/deleteAutoresponder}
}

