% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/batch.R
\name{v6_deleteBatch}
\alias{v6_deleteBatch}
\title{Deletes the specified Batch.}
\usage{
v6_deleteBatch(batch_id)
}
\arguments{
\item{batch_id}{(int) The ID of the Batch to delete.}
}
\value{
true on success.
}
\description{
Deletes the specified Batch.
}
\seealso{
\url{http://developers.vision6.com.au/3.0/method/deleteBatch}
}

