% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/folder.R
\name{v6_deleteFolder}
\alias{v6_deleteFolder}
\title{Deletes the specified Folder.}
\usage{
v6_deleteFolder(folder_id)
}
\arguments{
\item{folder_id}{(int) The ID of the Folder to delete.}
}
\value{
true on success.
}
\description{
Deletes the specified Folder.
}
\seealso{
\url{http://developers.vision6.com.au/3.0/method/deleteFolder}
}

