% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch.R
\name{v6_getAutoresponderBatchId}
\alias{v6_getAutoresponderBatchId}
\title{v6_getAutoresponderBatchId}
\usage{
v6_getAutoresponderBatchId()
}
\description{
v6_getAutoresponderBatchId
}
\seealso{
\url{http://developers.vision6.com.au/3.0/method/getAutoresponderBatchId}
}
