% Generated by roxygen2 (4.0.2): do not edit by hand
\name{v6_getBatchAutoDeactivated}
\alias{v6_getBatchAutoDeactivated}
\title{v6_getBatchAutoDeactivated}
\usage{
v6_getBatchAutoDeactivated()
}
\description{
v6_getBatchAutoDeactivated
}
\seealso{
\url{http://developers.vision6.com.au/3.0/method/getBatchAutoDeactivated}
}

