% Generated by roxygen2 (4.0.2): do not edit by hand
\name{v6_getBatchContactLinks}
\alias{v6_getBatchContactLinks}
\title{v6_getBatchContactLinks}
\usage{
v6_getBatchContactLinks()
}
\description{
v6_getBatchContactLinks
}
\seealso{
\url{http://developers.vision6.com.au/3.0/method/getBatchContactLinks}
}

