% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch.R
\name{v6_getBatchContactLinks}
\alias{v6_getBatchContactLinks}
\title{v6_getBatchContactLinks}
\usage{
v6_getBatchContactLinks(batch_id, send_id = 0, search_criteria = "[]",
  limit = 0, offset = 0, sort_by = "", sort_order = "")
}
\description{
v6_getBatchContactLinks
}
\seealso{
\url{http://developers.vision6.com.au/3.0/method/getBatchContactLinks}
}
