% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/batch.R
\name{v6_getBatchIdByQueueId}
\alias{v6_getBatchIdByQueueId}
\title{v6_getBatchIdByQueueId}
\usage{
v6_getBatchIdByQueueId(queue_id)
}
\description{
v6_getBatchIdByQueueId
}
\seealso{
\url{http://developers.vision6.com.au/3.0/method/getBatchIdByQueueId}
}

