% Generated by roxygen2 (4.0.2): do not edit by hand
\name{v6_getBatchLinkContacts}
\alias{v6_getBatchLinkContacts}
\title{v6_getBatchLinkContacts}
\usage{
v6_getBatchLinkContacts()
}
\description{
v6_getBatchLinkContacts
}
\seealso{
\url{http://developers.vision6.com.au/3.0/method/getBatchLinkContacts}
}

