% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch.R
\name{v6_getBatchLinkContacts}
\alias{v6_getBatchLinkContacts}
\title{v6_getBatchLinkContacts}
\usage{
v6_getBatchLinkContacts(batch_id, link, search_criteria = "[]", limit = 0,
  offset = 0, sort_by = "", sort_order = "")
}
\description{
v6_getBatchLinkContacts
}
\seealso{
\url{http://developers.vision6.com.au/3.0/method/getBatchLinkContacts}
}
