% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch.R
\name{v6_getBatchLinkStatistics}
\alias{v6_getBatchLinkStatistics}
\title{v6_getBatchLinkStatistics}
\usage{
v6_getBatchLinkStatistics(batch_id)
}
\description{
v6_getBatchLinkStatistics
}
\seealso{
\url{http://developers.vision6.com.au/3.0/method/getBatchLinkStatistics}
}
