% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/message.R
\name{v6_getMessageById}
\alias{v6_getMessageById}
\title{v6_getMessageById}
\usage{
v6_getMessageById()
}
\description{
v6_getMessageById
}
\seealso{
\url{http://developers.vision6.com.au/3.0/method/getMessageById}
}

