% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/message.R
\name{v6_getMessagesSentToContact}
\alias{v6_getMessagesSentToContact}
\title{v6_getMessagesSentToContact}
\usage{
v6_getMessagesSentToContact()
}
\description{
v6_getMessagesSentToContact
}
\seealso{
\url{http://developers.vision6.com.au/3.0/method/getMessagesSentToContact}
}
