% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/list.R
\name{v6_getTimezoneList}
\alias{v6_getTimezoneList}
\title{Returns the list of preferred Olson timezones used in the system.}
\usage{
v6_getTimezoneList()
}
\value{
a data frame of timezones with their names.
}
\description{
Returns the list of preferred Olson timezones used in the system.
}
\seealso{
\url{http://developers.vision6.com.au/3.0/method/getTimezoneList}
}

