% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/autoresponder.R
\name{v6_searchAutoresponders}
\alias{v6_searchAutoresponders}
\title{v6_searchAutoresponders}
\usage{
v6_searchAutoresponders()
}
\description{
v6_searchAutoresponders
}
\seealso{
\url{http://developers.vision6.com.au/3.0/method/searchAutoresponders}
}

